rework analytics provider and hooks
