+++ packages/york-analytics/README.MD
@@ -1,18 +1,33 @@
 `york-analytics` — это набор компонентов и утилит, призванный упростить сбор аналитических данных.
 
+###  TL / DR
+Для минимальной настройки необходимо добавить в корень приложения `RootAnalyticsProvider`. После этого все `Button`, `Screen` и `Page` начнут отправлять события автоматически.
+
+```js static
+import React from 'react'
+import { RootAnalyticsProvider } from '@qlean/york-analytics'
+
+const App = () => (
+  <RootAnalyticsProvider appId="exampleApp" trackEvent={console.log}>
+    // children
+  </RootAnalyticsProvider>
+)
+```
+
 ### Как это работает?
 
 Вся система работает через React Context. Ключевым модулем является `AnalyticsProvider`, который передает дочерним элементам аналитический контекст. Провайдеры можно бесконечно вкладывать друг в друга, создавая новые контексты, при этом дочерние элементы будут брать информацию из ближайшего к ним провайдера.
 
 ### Как выглядит аналитическое событие?
 
-У любого события есть как минимум 3 свойства: `category`, `label` и `action`. Но к событию можно добавить любое количество других данных.
+У любого события есть как минимум 3 свойства: `category`, `label` и `action`. По умолчанию к каждому событию добавляется свойство `analyticsRoute`, с помощью которого можно понять где именнов приложении произошло событие. К событию можно добавить любое количество других данных.
 
 ```js static
 {
   category: String, // Категория события, как правило совпадает с названием экрана или страницы
   label: String, // Название элемента, отправляющего событие
   action: String, // Название действия
+  analyticsRoute: String // Генерируется автоматически из иерархии провайдеров
   ...analyticsData: Object, // Дополнительные данные, которые нужно приложить к событию
 }
 
@@ -36,12 +51,9 @@
 Стандартные названия действий содержатся в модуле `eventActionTypes`.
 ```
 
-Событие можно дополнить свойством `analyticsRoute`, которое позволяет понять, где именно в приложении совершилось действие. Оно генерируется автоматически и хранится в контексте. Компоненты `Button`, `Page` и `Header` из `york-web`, а так же `Button` и `Screen` из `york-react-native` автоматически добавляют `analyticsRoute` ко всем событиям, которые они отправляют.
-Пример: `york/examplePage`
-
 ### Как события отправляются на сервер?
 
-В `AnalyticsProvider` можно передать проп `trackEvent(event: Object): Function` - функцию, которая принимает событие и передает его в транспорт. Транспорт может быть разный у разных приложений. `trackEvent` можно передать в любой провайдер (например если в какой-то части приложения нужно использовать другой трекинг), но необходимо сделать это как минимум в самом первом. После передачи этой функции, все дочерние провайдеры и подключенные компоненты получат к ней доступ через контекст.
+В `RootAnalyticsProvider` передается проп `trackEvent(event: Object): Function` - функцию, которая принимает событие и передает его в транспорт. Транспорт может быть разный у разных приложений. После передачи этой функции, все дочерние провайдеры и подключенные компоненты получат к ней доступ через контекст.
 
 ### Совместимые компоненты
 
@@ -49,36 +61,32 @@
 
 #### york-web
 
-`Page` — автоматически создает аналитический контекст и отправляет событие `mount` при первом рендере. Категория события равна свойству `name`
+`Page` — автоматически создает аналитический контекст и отправляет событие `pageView`. Категория события равна свойству `name`
 `Button` — автоматически отправляет событие `click` при нажатии. Категория события берется из ближайшего провайдера.
 `Header` — автоматически отправляет события при клике на элементы, вызывающие коллбэки. Категория: `header`.
 
 #### york-react-native
 
-`Screen` — автоматически создает аналитический контекст и отправляет событие `mount` при первом рендере. Категория события равна свойству `name`
-`Button` — автоматически отправляет событие `click` при нажатии. Категория события берется из ближайшего провайдера.
+`Screen` — автоматически создает аналитический контекст и отправляет событие `screenView`. Категория события равна свойству `name`
+`Button` — автоматически отправляет событие `press` при нажатии. Категория события берется из ближайшего провайдера.
 
 ### Добавление трекинга к кастомным компонентам
 
 ```js static
 import React, { useContext } from 'react'
 import PropTypes from 'prop-types'
-import { eventActionTypes, AnalyticsContext } from '@qlean/york-analytics'
+import { eventActionTypes, useAnalytics } from '@qlean/york-analytics'
 
 const CustomButton = ({ onClick, analyticsData, ...rest }) => {
-  const analyticsContext = useContext(AnalyticsContext)
+  const { trackEvent, category } = useAnalytics()
 
   const handleClick = e => {
-    if (analyticsContext) {
-      const { trackEvent, category, analyticsRoute } = analyticsContext
-      trackEvent({
-        category,
-        label: name,
-        action: eventActionTypes.click,
-        analyticsRoute,
-        ...analyticsData,
-      })
-    }
+    trackEvent({
+      category,
+      label: name,
+      action: eventActionTypes.click,
+      ...analyticsData,
+    })
     onClick(e)
   }
 
@@ -95,60 +103,4 @@ CustomButton.propTypes = {
 }
 
 export default CustomButton
-```
-
-### Трекинг ссылок
-
-Ссылки в разных приложениях могут вести себя по разному, поэтому на данный момент в `york-analytics` нет универсального решения для трекинга нажатия на них.
-
-Вот пример реализации для ссылки с трекингом:
-
-```js static
-import React, { useContext, useEffect, useState } from 'react'
-import PropTypes from 'prop-types'
-import { Link as YorkLink } from '@qlean/york-web'
-import { eventActionTypes, AnalyticsContext } from '@qlean/york-analytics'
-
-// Функция, которая преобразует url. В идеале является частью модуля транспорта.
-import { getAnalyticsUrl } from 'modules/analytics/utils'
-
-const Link = ({ href, name, analyticsData, ...rest }) => {
-  const [url, setUrl] = useState(href)
-  const analyticsContext = useContext(AnalyticsContext)
-
-  useEffect(() => {
-    if (analyticsContext) {
-      const externalPattern = new RegExp('^(?:[a-z]+:)?//', 'i')
-      const isExternalLink = externalPattern.test(href)
-      const { category, analyticsRoute } = analyticsContext
-      setUrl(
-        getAnalyticsUrl({
-          href: isExternalLink ? href : `${window.location.origin}${href}`,
-          category,
-          label: name,
-          action: eventActionTypes.click,
-          redirectUrl: 'https://anlt.cloud.qlean.ru/collect',
-          analyticsRoute,
-          destination: href,
-          ...analyticsData,
-        }),
-      )
-    }
-  }, [href, name, analyticsContext, analyticsData])
-
-  return <YorkLink href={url} name={name} {...rest} />
-}
-
-Link.defaultProps = {
-  analyticsData: {},
-}
-
-Link.propTypes = {
-  href: PropTypes.string.isRequired,
-  name: PropTypes.string.isRequired,
-  // eslint-disable-next-line react/forbid-prop-types
-  analyticsData: PropTypes.object,
-}
-
-export default Link
 ```
