+++ packages/york-analytics/docs/context.md
@@ -5,13 +5,12 @@ import { AnalyticsContext, eventActionTypes } from '@qlean/york-analytics'
 
 const ExampleComponent = () => (
   <AnalyticsContext.Consumer>
-    {({ trackEvent, category, analyticsRoute }) => (
+    {({ trackEvent, category }) => (
       <button
         onClick={trackEvent({
           category,
           label: 'doExampleAction',
           action: eventActionTypes.click,
-          analyticsRoute,
         })}
       >
         ExampleButton
