+++ packages/york-analytics/src/constants.js
@@ -1,6 +1,8 @@
 export const eventActionTypes = {
-  mount: 'mount',
   click: 'click',
+  pageView: 'pageView',
+  mount: 'mount',
   press: 'press',
+  screenView: 'screenView',
   submit: 'submit',
 }
