+++ packages/york-analytics/src/index.js
@@ -1,4 +1,7 @@
-export * from './constants'
-export { default as AnalyticsContext } from './context'
 export { default as AnalyticsProvider } from './components/AnalyticsProvider'
-export * from './utils/hooks'
+export {
+  default as RootAnalyticsProvider,
+} from './components/RootAnalyticsProvider'
+export * from './constants'
+export * from './context'
+export * from './utils'
