+++ packages/york-analytics/src/utils/index.js
@@ -1,0 +1,2 @@
+export { default as useAnalytics } from './hooks/useAnalytics'
+export { default as useViewTracking } from './hooks/useViewTracking'
