+++ packages/york-react-native/src/components/Button/index.js
@@ -168,12 +168,11 @@ const Button = ({
 
   const handlePress = e => {
     if (analyticsContext) {
-      const { trackEvent, category, analyticsRoute } = analyticsContext
+      const { trackEvent, category } = analyticsContext
       trackEvent({
         category,
         label: name,
         action: eventActionTypes.press,
-        analyticsRoute,
         ...analyticsData,
       })
     }
