+++ packages/york-web/src/components/primitive/Page/index.js
@@ -1,92 +1,16 @@
-import React, { useEffect, useRef } from 'react'
+import React from 'react'
 import PropTypes from 'prop-types'
 
-import {
-  AnalyticsProvider,
-  eventActionTypes,
-  AnalyticsContext,
-} from '@qlean/york-analytics'
-
-const usePrevious = value => {
-  const ref = useRef()
-
-  useEffect(() => {
-    ref.current = value
-  }, [value])
-
-  return ref.current
-}
-
-const shouldSendEvent = data =>
-  Object.values(data).filter(item => Boolean(item)).length ===
-  Object.keys(data).length
+import { AnalyticsProvider, useViewTracking } from '@qlean/york-analytics'
 
 const Page = ({ name, analyticsData, children }) => {
-  console.log('render', name, analyticsData)
-
-  const prev = usePrevious(analyticsData)
-  const { trackEvent } = useAnalytics()
-
-  useEffect(() => {
-    if (!analyticsData) {
-      trackEvent({
-        label: name,
-        action: eventActionTypes.pageView,
-      })
-    }
-
-    // Мы можем спокойно использовать сравнение строк потому что
-    // подразумевается что analyticsData - это объект, формируемый вручную,
-    // с плоской структурой, и он все равно конвертируется в JSON перед отправкой.
-
-    const hasDataUpdated =
-      prev && JSON.stringify(analyticsData) !== JSON.stringify(prev)
-
-    if (analyticsData && shouldSendEvent(analyticsData) && hasDataUpdated) {
-      trackEvent({
-        label: name,
-        action: 'mount',
-        ...analyticsData,
-      })
-    }
-  }, [name, analyticsData, trackEvent, prev])
-
+  useViewTracking({ name, analyticsData })
   return <AnalyticsProvider category={name}>{children}</AnalyticsProvider>
 }
 
-export default Page
-
-// import React, { useEffect, useContext } from 'react'
-// import PropTypes from 'prop-types'
-// import {
-//   AnalyticsProvider,
-//   eventActionTypes,
-//   AnalyticsContext,
-// } from '@qlean/york-analytics'
-
-// /** Компонет, отвечающий за рендер страницы. `Page` автоматически создает новый контекст для аналитики (см. york-analytics) */
-// const Page = ({ name, analyticsData, children }) => {
-//   const analyticsContext = useContext(AnalyticsContext)
-
-//   useEffect(() => {
-//     if (analyticsContext) {
-//       const { trackEvent, analyticsRoute } = analyticsContext
-//       trackEvent({
-//         label: name,
-//         category: name,
-//         action: eventActionTypes.mount,
-//         analyticsRoute,
-//         ...analyticsData,
-//       })
-//     }
-//   }, [analyticsContext, analyticsData, name])
-
-//   return analyticsContext ? (
-//     <AnalyticsProvider category={name}>{children}</AnalyticsProvider>
-//   ) : (
-//     children
-//   )
-// }
+Page.defaultProps = {
+  analyticsData: null,
+}
 
 Page.propTypes = {
   /** Имя страницы. Используется для аналитики */
@@ -98,8 +22,4 @@ Page.propTypes = {
   children: PropTypes.node.isRequired,
 }
 
-Page.defaultProps = {
-  analyticsData: {},
-}
-
-// export default Page
+export default Page
