+++ packages/york-web/src/components/simple/Button/index.js
@@ -218,12 +218,11 @@ function Button({
   const handleClick = e => {
     if (onClick) {
       if (analyticsContext) {
-        const { trackEvent, category, analyticsRoute } = analyticsContext
+        const { trackEvent, category } = analyticsContext
         trackEvent({
           category,
           label: name,
           action: eventActionTypes.click,
-          analyticsRoute,
           ...analyticsData,
         })
       }
